requirements.txt
langchain
langchain-core
langchain-community
pypdf
pymupdf
ipykernel
sentence-transformers
pinecone
chromadb
faiss-cpu
langchain-groq
python-dotenv
streamlit